pipeliner: New element for fast thread offloading
This element is a fast version of ThreadSafeQueue->Unqueue, as it does not involve the heavy push->pull notification mechanism and
rely on a simple task rescheduling limited by a sleepiness threshold.

Also introduce the ring.hh library, featuring a simple single consumer single producer queue. Storage cannot be used for pipeliner as there is multiple storage queue, one per thread to avoid locking.

The multithread.hh library give helpers for multithread management, for now it is just a per_thread<T> structure to duplicate T in a per-thread way. Unlike thread-local storage (__thread), value can be accessed by other thread. When this method can be used instead of software or hardware memory locking it is probably the faster way to deal with thread contention.
A good exemple is statistic collecting in pipeliner.cc : instead of having multiple thread incrementing the same variable, increment a per-thread variable and do the sum when you need to actually read the statistics. The obvious downside is a big loss of memory, but who needs that nowadays?