ToDevice(.k) : Set queue_mapping to 0
Somehow, even with only one queue in RX this field is dangling.
Fix #359. In ixgbe and i40e that lead to accessing an inexisting queue.

Proper support would be to initialize the device with multiple queues,
or us multiple ToDevice with a QUEUE argument and use one queue per
thread. But that's for another day...